Restore Tesseract playground build
