+++ src/LM.App.Wpf/LM.App.Wpf.csproj
@@ -21,6 +21,7 @@
     <PackageReference Include="Microsoft.Extensions.Hosting" />
     <PackageReference Include="CommunityToolkit.Mvvm" />
     <PackageReference Include="Microsoft.Xaml.Behaviors.Wpf" />
+    <PackageReference Include="Tesseract" />
     <PackageReference Include="Tabula" />
   </ItemGroup>
